Move node properties testing into a separate file
The single converter_test file was becoming extremely cluttered and
hard to work with. I also fixed test_transferred_properties to use the
new instance method instead of a class method
